requirements.txt
pytest
selenium
allure-pytest
selene==2.0.0rc9